Create CategoryTreeRepository and move dedicated methods in it
