+++ src/Akeneo/Component/Classification/Repository/CategoryRepositoryInterface.php
@@ -40,15 +40,6 @@ interface CategoryRepositoryInterface extends
     public function getCategoriesByCodes(array $categoryCodes = []);
 
     /**
-     * Get a tree filled with children and their parents
-     *
-     * @param array $parentsIds parent ids
-     *
-     * @return array
-     */
-    public function getTreeFromParents(array $parentsIds);
-
-    /**
      * Shortcut to get all children ids
      *
      * @param CategoryInterface $parent      the parent
@@ -78,96 +69,9 @@ interface CategoryRepositoryInterface extends
     public function getCategoryIdsByCodes(array $codes);
 
     /**
-     * Get children from a parent id
-     *
-     * @param int $parentId
-     *
-     * @return ArrayCollection
-     */
-    public function getChildrenByParentId($parentId);
-
-    /**
-     * @param CategoryInterface $parent
-     * @param array             $grantedCategoryIds
-     *
-     * @return array
-     */
-    public function getChildrenGrantedByParentId(CategoryInterface $parent, array $grantedCategoryIds = []);
-
-    /**
-     * Get children tree from a parent id.
-     * If the select node id is provided, the tree will be returned
-     * down to the node specified by select node id. Otherwise, the
-     * whole tree will be returned
-     *
-     * @param int  $parentId
-     * @param bool $selectNodeId
-     *
-     * @return array
-     */
-    public function getChildrenTreeByParentId($parentId, $selectNodeId = false, array $grantedCategoryIds = []);
-
-    /**
-     * Based on the Gedmo\Tree\RepositoryUtils\buildTreeArray, but with
-     * keeping the node as object and able to manage nodes in different branches
-     * (the original implementation works with only depth and associate all
-     * nodes of depth D+1 to the last node of depth D.)
-     *
-     * @param array $nodes Must be sorted by increasing depth
-     *
-     * @return array
-     */
-    public function buildTreeNode(array $nodes);
-
-    /**
-     * Get the Tree path of Nodes by given $node
-     *
-     * @param object $node
-     *
-     * @return array - list of Nodes in path
-     */
-    public function getPath($node);
-
-    /**
-     * @return array
-     */
-    public function getTrees();
-
-    /**
-     * Get trees of granted categories
-     *
-     * @param array $grantedCategoryIds
-     *
-     * @return array
-     */
-    public function getGrantedTrees(array $grantedCategoryIds = []);
-
-    /**
-     * Check if a parent node is an ancestor of a child node
-     *
-     * @param CategoryInterface $parentNode
-     * @param CategoryInterface $childNode
-     *
-     * @return bool
-     */
-    public function isAncestor(CategoryInterface $parentNode, CategoryInterface $childNode);
-
-    /**
      * Return the categories sorted by tree and ordered
      *
      * @return array
      */
     public function getOrderedAndSortedByTreeCategories();
-
-    /**
-     * Provides a tree filled up to the categories provided, with all their ancestors
-     * and ancestors sibligns are filled too, in order to be able to display the tree
-     * directly without loading other data.
-     *
-     * @param CategoryInterface $root       Tree root category
-     * @param Collection        $categories Collection of categories
-     *
-     * @return array Multi-dimensional array representing the tree
-     */
-    public function getFilledTree(CategoryInterface $root, Collection $categories);
 }
