+++ src/Pim/Bundle/EnrichBundle/Resources/config/repositories.yml
@@ -20,6 +20,7 @@ parameters:
     pim_enrich.repository.job_instance.class:            Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobInstanceRepository
     pim_enrich.repository.job_execution.class:           Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobExecutionRepository
     pim_enrich.repository.job_tracker.class:             Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobTrackerRepository
+    pim_enrich.repository.category_tree.class:           Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\CategoryTreeRepository
 
 services:
     pim_enrich.repository.sequential_edit:
@@ -139,3 +140,9 @@ services:
         arguments:
             - '@doctrine.orm.entity_manager'
             - '%pim_catalog.entity.attribute_option.class%'
+
+    pim_enrich.repository.category_tree:
+        class: '%pim_enrich.repository.category_tree.class%'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.category.class%'
